# backend/ShuKnow.WebAPI/Dto/Folders/FolderTreeNodeDto.cs
namespace ShuKnow.WebAPI.Dto.Folders;

public record FolderTreeNodeDto(
    Guid Id,
    string Name,
    string Description,
    string? Emoji,
    int SortOrder,
    int FileCount,
    IReadOnlyList<FolderTreeNodeDto> Children);