# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShuKnow API
  description: REST API for the ShuKnow AI-powered file organization platform.
  version: 1.0.0
  contact:
    name: ShuKnow Team
servers:
  - url: https://localhost:5432
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/folders/tree:
    get:
      operationId: getFolderTree
      tags: [Folders]
      summary: Get the full folder tree
      description: |
        Returns the complete hierarchical folder tree for the authenticated user.
        Each node contains children recursively and a file count. Designed for
        rendering the sidebar. Not paginated — unlikely to exceed a few hundred
        nodes per user.
      responses:
        '200':
          description: Full folder tree.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderTreeNodeDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders:
    get:
      operationId: listFolders
      tags: [Folders]
      summary: List folders at a given level
      description: |
        Returns a flat list of folders at a specific hierarchy level.
        When `parentId` is omitted or null, returns root-level folders.
        Lightweight alternative to the full tree endpoint.
      parameters:
        - name: parentId
          in: query
          required: false
          description: Parent folder ID. Omit or set to null for root folders.
          schema:
            type: string
            format: uuid
            nullable: true
      responses:
        '200':
          description: List of folders at the specified level.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createFolder
      tags: [Folders]
      summary: Create a new folder
      description: |
        Creates a folder in the hierarchy. Validates name uniqueness within the
        parent scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A folder with this name already exists in the target scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/folders/{folderId}:
    parameters:
      - $ref: '#/components/parameters/FolderId'

    get:
      operationId: getFolder
      tags: [Folders]
      summary: Get a single folder
      description: |
        Returns folder metadata including a `path` breadcrumb array showing
        the full hierarchy from root to this folder.
      responses:
        '200':
          description: Folder details with breadcrumb path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateFolder
      tags: [Folders]
      summary: Update folder name and/or description
      description: Validates name uniqueness within the parent scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolderRequest'
      responses:
        '200':
          description: Folder updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A sibling folder with this name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      operationId: deleteFolder
      tags: [Folders]
      summary: Delete a folder
      description: |
        Deletes a folder. If `recursive=true`, all children folders and files
        are deleted. If `recursive=false` (default) and the folder is non-empty,
        returns 409 Conflict.
      parameters:
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, delete all nested folders and files.
      responses:
        '204':
          description: Folder deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Folder is non-empty and `recursive` is false.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/folders/{folderId}/move:
    parameters:
      - $ref: '#/components/parameters/FolderId'

    patch:
      operationId: moveFolder
      tags: [Folders]
      summary: Move a folder to a new parent
      description: |
        Moves a folder to a different parent (or to root if `newParentFolderId`
        is null). Validates that the move does not create a cycle and that the
        folder name is unique in the target scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFolderRequest'
      responses:
        '200':
          description: Folder moved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderDto'
        '400':
          description: Move would create a cycle or other validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Name conflict in target parent scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/folders/{folderId}/reorder:
    parameters:
      - $ref: '#/components/parameters/FolderId'

    patch:
      operationId: reorderFolder
      tags: [Folders]
      summary: Change folder sort position among siblings
      description: |
        Sets the target 0-based position for this folder among its siblings.
        The server re-indexes all sibling sort orders accordingly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderFolderRequest'
      responses:
        '204':
          description: Sort order updated.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/folders/{folderId}/children:
    parameters:
      - $ref: '#/components/parameters/FolderId'

    get:
      operationId: getFolderChildren
      tags: [Folders]
      summary: List direct child folders
      description: |
        Returns the immediate child folders of the specified folder.
        Useful for lazy-loading expanded tree nodes.
      responses:
        '200':
          description: Direct child folders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/folders/{folderId}/files:
    parameters:
      - $ref: '#/components/parameters/FolderId'

    get:
      operationId: listFolderFiles
      tags: [Folders]
      summary: List files in a folder (paginated)
      description: Offset-based pagination. Default page size is 50.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated list of files.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedFileResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: uploadFile
      tags: [Files]
      summary: Upload a file into a folder
      description: |
        Uploads a file via `multipart/form-data`. If `name` is omitted the
        original filename is used. Validates name uniqueness within the folder.
        File size limit is configurable (e.g. 10 MB for MVP).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload.
                name:
                  type: string
                  description: Override display name. Defaults to original filename.
                description:
                  type: string
                  description: Optional description for the file.
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: File uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Folder not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: A file with this name already exists in the folder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: File exceeds the maximum allowed size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ──────────────────────────────────────────────
  # Files
  # ──────────────────────────────────────────────
  /api/files/{fileId}:
    parameters:
      - $ref: '#/components/parameters/FileId'

    get:
      operationId: getFile
      tags: [Files]
      summary: Get file metadata
      description: Returns file metadata including name, description, content type, and size.
      responses:
        '200':
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateFile
      tags: [Files]
      summary: Update file name and/or description
      description: Validates name uniqueness within the file's current folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        '200':
          description: File metadata updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A file with this name already exists in the folder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Permanently delete a file
      responses:
        '204':
          description: File deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in the `Authorization` header.

    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token in an HTTP-only cookie (`SameSite=None; Secure`).

  # ──────────────────────────────────────────────
  # Reusable Parameters
  # ──────────────────────────────────────────────
  parameters:
    FolderId:
      name: folderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the folder.
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the file.
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Number of items per page.

  # ──────────────────────────────────────────────
  # Reusable Responses
  # ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required. JWT is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationError:
      description: One or more validation errors occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
...
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: Invoices
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Monthly invoices and receipts
        parentFolderId:
          type: string
          format: uuid
          nullable: true
          description: Parent folder ID. Null or omitted for root-level.
        emoji:
          type: string
          nullable: true
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
        emoji:
          type: string
          nullable: true
    MoveFolderRequest:
      type: object
      properties:
        newParentFolderId:
          type: string
          format: uuid
          nullable: true
          description: Target parent folder ID. Null to move to root.
    ReorderFolderRequest:
      type: object
      required: [position]
      properties:
        position:
          type: integer
          minimum: 0
          description: 0-based target index among siblings.
          example: 2
    FolderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Invoices
        description:
          type: string
          example: Monthly invoices and receipts
        emoji:
          type: string
          nullable: true
        parentFolderId:
          type: string
          format: uuid
          nullable: true
        sortOrder:
          type: integer
          example: 0
        fileCount:
          type: integer
          example: 12
        hasChildren:
          type: boolean
          example: true
        path:
          type: array
          nullable: true
          items:
            type: string
          description: |
            Breadcrumb path from root to this folder. Populated on
            `GET /api/folders/{folderId}`.
          example: ["Documents", "Finance", "Invoices"]
    FolderTreeNodeDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Documents
        description:
          type: string
        emoji:
          type: string
          nullable: true
        sortOrder:
          type: integer
        fileCount:
          type: integer
        children:
          type: array
          items:
            $ref: '#/components/schemas/FolderTreeNodeDto'

    # ── Files ─────────────────────────────────
    UpdateFileRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
...
    FileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
        folderName:
          type: string
          example: Invoices
        name:
          type: string
          example: invoice-2024-01.pdf
        description:
          type: string
          example: January 2024 electricity bill
        contentType:
          type: string
          example: application/pdf
        sizeBytes:
          type: integer
          format: int64
          example: 204800
        sortOrder:
          type: integer
          description: Position within the parent folder.
          example: 0
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp.

    # ── Chat ──────────────────────────────────
...
    PagedFileResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FileDto'
        totalCount:
          type: integer
          example: 128
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 50
        hasNextPage:
          type: boolean
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details.
      properties:
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7807
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Object with ID '...' was not found.
        instance:
          type: string
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                name: ["The Name field is required."]
                description: ["Description must not exceed 1000 characters."]